requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.27
SQLAlchemy-Utils==0.41.1
typing_extensions==4.10.0
validators==0.22.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
